requirements.txt
gradio
pandas
numpy
python-dotenv
langchain
langchain-chroma
langchain-huggingface
sentence-transformers
chromadb
torch
